+++ R/gg_vistime.R
@@ -3,19 +3,19 @@
 #' Provide a data frame with event data to create a static timeline plot.
 #' Simplest drawable dataframe can have columns `event` and `start`.
 #'
-#' @param data \code{data.frame} that contains the data to be visualised
-#' @param events (optional, character) the column name in \code{data} that contains event
+#' @param data \code{data.frame} that contains the data to be visualized
+#' @param col.event (optional, character) the column name in \code{data} that contains event
 #'   names. Default: \emph{event}.
-#' @param start (optional, character) the column name in \code{data} that contains start
+#' @param col.start (optional, character) the column name in \code{data} that contains start
 #'   dates. Default: \emph{start}.
-#' @param end (optional, character) the column name in \code{data} that contains end dates.
+#' @param col.end (optional, character) the column name in \code{data} that contains end dates.
 #'   Default: \emph{end}.
-#' @param groups (optional, character) the column name in \code{data} to be used for
+#' @param col.group (optional, character) the column name in \code{data} to be used for
 #'   grouping. Default: \emph{group}.
-#' @param colors (optional, character) the column name in \code{data} that contains colors
+#' @param col.color (optional, character) the column name in \code{data} that contains colors
 #'   for events. Default: \emph{color}, if not present, colors are chosen via
 #'   \code{RColorBrewer}.
-#' @param fontcolors (optional, character) the column name in \code{data} that contains the
+#' @param col.fontcolor (optional, character) the column name in \code{data} that contains the
 #'   font color for event labels. Default: \emph{fontcolor}, if not present,
 #'   color will be black.
 #' @param optimize_y (optional, logical) distribute events on y-axis by smart heuristic (default),
@@ -26,7 +26,8 @@
 #'   Default: \code{NULL}.
 #' @param show_labels (optional, boolean) choose whether or not event labels shall be
 #'   visible. Default: \code{TRUE}.
-#' @param background_lines (optional, integer) the number of vertical lines to draw in the background to demonstrate structure (default: 10). Less means more memory-efficient plot.
+#' @param background_lines (optional, integer) the number of vertical lines to draw in the background to demonstrate structure (default: heuristic).
+#' @param ... for deprecated arguments up to vistime 1.1.0 (like events, colors, ...)
 #' @export
 #' @return \code{gg_vistime} returns an object of class \code{gg} and \code{ggplot}.
 #' @examples
@@ -39,15 +40,47 @@
 #'   color = c("#cbb69d", "#603913", "#c69c6e")
 #' )
 #'
-#' gg_vistime(pres, events = "Position", groups = "Name", title = "Presidents of the USA")
+#' gg_vistime(pres, col.event = "Position", col.group = "Name", title = "Presidents of the USA")
 
-gg_vistime <- function(data, events = "event", start = "start", end = "end", groups = "group",
-                       colors = "color", fontcolors = "fontcolor", optimize_y = TRUE,
-                       linewidth = NULL, title = NULL, show_labels = TRUE, background_lines = 10) {
+gg_vistime <- function(data, col.event = "event", col.start = "start", col.end = "end", col.group = "group",
+                       col.color = "color", col.fontcolor = "fontcolor", optimize_y = TRUE,
+                       linewidth = NULL, title = NULL, show_labels = TRUE, background_lines = NULL, ...) {
 
-  data$tooltip <- ""
-  data <- validate_input(data, start, end, events, groups, tooltips = "tooltip", optimize_y, linewidth, title, show_labels, background_lines)
-  cleaned_dat <- vistime_data(data, events, start, end, groups, colors, fontcolors, tooltips = "tooltip", optimize_y)
+  .dots = list(...)
+
+  if("events" %in% names(.dots)){
+    .Deprecated(new = "col.event", old = "events")
+    col.event = .dots$events
+  }
+  if("start" %in% names(.dots)){
+    .Deprecated(new = "col.start", old = "start")
+    col.start = .dots$start
+  }
+  if("end" %in% names(.dots)){
+    .Deprecated(new = "col.end", old = "end")
+    col.end = .dots$end
+  }
+  if("groups" %in% names(.dots)){
+    .Deprecated(new = "col.group", old = "groups")
+    col.group = .dots$groups
+  }
+  if("colors" %in% names(.dots)){
+    .Deprecated(new = "col.color", old = "colors")
+    col.color = .dots$colors
+  }
+  if("fontcolors" %in% names(.dots)){
+    .Deprecated(new = "col.fontcolor", old = "fontcolors")
+    col.fontcolor = .dots$fontcolors
+  }
+  if("lineInterval" %in% names(.dots)){
+    .Deprecated(new = "background_lines", old = "lineInterval")
+  }
+  if("showLabels" %in% names(.dots)){
+    .Deprecated(new = "show_labels", old = "showLabels")
+  }
+
+  data <- validate_input(data, col.event, col.start, col.end, col.group, col.tooltip = NULL, optimize_y, linewidth, title, show_labels, background_lines)
+  cleaned_dat <- vistime_data(data, col.event, col.start, col.end, col.group, col.color, col.fontcolor, col.tooltip = "tooltip", optimize_y)
 
   total <- plot_ggplot(cleaned_dat, linewidth, title, show_labels, background_lines)