+++ tests/testthat/test_set_colors.R
@@ -1,11 +1,11 @@
 # preparations
-events <- "event"
-start <- "start"
-end <- "end"
-groups <- "group"
-colors <- "color"
-fontcolors <- "fontcolor"
-tooltips <- "tooltip"
+col.event <- "event"
+col.start <- "start"
+col.end <- "end"
+col.group <- "group"
+col.color <- "color"
+col.fontcolor <- "fontcolor"
+col.tooltip <- "tooltip"
 optimize_y <- TRUE
 linewidth <- NULL
 title <- NULL
@@ -17,24 +17,24 @@ data <- data.frame(
   FARBE = c("red", "green"),
   stringsAsFactors = FALSE
 )
-data <- vistime:::validate_input(data, start, end, events, groups, tooltips, optimize_y, linewidth, title, show_labels, background_lines)
+data <- validate_input(data, col.start, col.end, col.event, col.group, col.tooltip, optimize_y, linewidth, title, show_labels, background_lines)
 
 
 test_that("color columns not existing", {
-  result <- vistime:::set_colors(data, eventcolor_column = "NOTEXISTING", fontcolor_column = "NOTEXISTING")
+  result <- set_colors(data, eventcolor_column = "NOTEXISTING", fontcolor_column = "NOTEXISTING")
   expect_true(all(c("col", "fontcol") %in% names(result)))
   expect_true(all(result$fontcol == "black"))
   expect_equal(result$col, c("#8DD3C7", "#FFFFB3", "#BEBADA", "#FB8072"))
 })
 
 test_that("eventcolor column existing", {
-  result <- vistime:::set_colors(data, eventcolor_column = "FARBE", fontcolor_column = "NOTEXISTING")
+  result <- set_colors(data, eventcolor_column = "FARBE", fontcolor_column = "NOTEXISTING")
   expect_true(all(c("col", "fontcol") %in% names(result)))
   expect_equal(result$col, c("red", "green", "red", "green"))
 })
 
 test_that("fontcolor column existing", {
-  result <- vistime:::set_colors(data, eventcolor_column = "NOTEXISTING", fontcolor_column = "FARBE")
+  result <- set_colors(data, eventcolor_column = "NOTEXISTING", fontcolor_column = "FARBE")
   expect_true(all(c("col", "fontcol") %in% names(result)))
   expect_true(all(result$fontcol == result$FARBE))
   expect_equal(result$fontcol, c("red", "green", "red", "green"))
@@ -42,7 +42,7 @@ test_that("fontcolor column existing", {
 
 
 test_that("trim whitespaces", {
-  fixed <- vistime:::set_colors(data.frame(event = "Event1", start = "2014-01-01", color = "   #676767"),
+  fixed <- set_colors(data.frame(event = "Event1", start = "2014-01-01", color = "   #676767"),
                                         eventcolor_column = "color", fontcolor_column = "NOTHING")[, c("col", "fontcol")]
 
   expect_equivalent(fixed, lapply(fixed, trimws))