+++ tests/testthat/test_set_order.R
@@ -1,21 +1,21 @@
 
  prepare_data <- function(dat){
-  events <- "event"
-  start <- "start"
-  end <- "end"
-  groups <- "group"
-  colors <- "color"
-  fontcolors <- "fontcolor"
-  tooltips <- "tooltip"
+  col.event <- "event"
+  col.start <- "start"
+  col.end <- "end"
+  col.group <- "group"
+  col.color <- "color"
+  col.fontcolor <- "fontcolor"
+  col.tooltip <- "tooltip"
   optimize_y <- TRUE
   linewidth <- NULL
   title <- NULL
   show_labels <- TRUE
   background_lines <- 11
 
-  dat <- vistime:::validate_input(dat, start, end, events, groups, tooltips, optimize_y, linewidth, title, show_labels, background_lines)
-  dat <- vistime:::set_colors(dat, colors, fontcolors)
-  dat <- vistime:::fix_columns(dat, events, start, end, groups, tooltips)
+  dat <- validate_input(dat, col.start, col.end, col.event, col.group, col.tooltip, optimize_y, linewidth, title, show_labels, background_lines)
+  dat <- set_colors(dat, col.color, col.fontcolor)
+  dat <- fix_columns(dat, col.event, col.start, col.end, col.group, col.tooltip)
   return(dat)
 }
 
@@ -29,18 +29,18 @@ dat <- prepare_data(dat)
 
 test_that("ranges: no groups yield 1 subplot", {
   dat <- dat[dat$start != dat$end]
-  expect_equal(1, unique(vistime:::set_order(dat)$subplot))
+  expect_equal(1, unique(set_order(dat)$subplot))
 })
 
 test_that("one subplot per group", {
   dat$group <- c(1,1,2)
-  result <- vistime:::set_order(dat)
+  result <- set_order(dat)
   expect_equal(as.character(result$group), as.character(result$subplot))
 })
 
 test_that("if range and event in same group, different subplots", {
   dat$group <- c(1, 2, 1)
-  result <- vistime:::set_order(dat)
+  result <- set_order(dat)
 
   expect_gt(length(result[result$group == 1, "subplot"]), 1) # Group 1 must have more than 1 subplot
   expect_equal(unique(result$subplot[result$group == 1]), 1) # Group 1 must have a single subplot
@@ -56,7 +56,7 @@ test_that("subplot order is same a input order", {
   d$target_subplot <- c(1,2)
 
   dat <- prepare_data(d)
-  expect_equal(vistime:::set_order(dat)$subplot, d$target_subplot)
+  expect_equal(set_order(dat)$subplot, d$target_subplot)
 })
 
 
@@ -69,8 +69,8 @@ test_that("More than 9 groups are handled nicely", {
   )
 
   dat <- prepare_data(dat)
-  expect_equal(vistime:::set_order(dat)$group, factor(dat$group, levels = unique(dat$group)))
-  expect_equal(vistime:::set_order(dat)$subplot, 1:15)
+  expect_equal(set_order(dat)$group, factor(dat$group, levels = unique(dat$group)))
+  expect_equal(set_order(dat)$subplot, 1:15)
 })
 
 test_that("Groups are gathered together", {
@@ -81,6 +81,6 @@ test_that("Groups are gathered together", {
   )
 
   dat <- prepare_data(dat)
-  expect_equal(as.integer(as.character(vistime:::set_order(dat)$group)), c(2,2,1,1,5))
-  expect_equal(vistime:::set_order(dat)$subplot, c(1,1,2,2,3))
+  expect_equal(as.integer(as.character(set_order(dat)$group)), c(2,2,1,1,5))
+  expect_equal(set_order(dat)$subplot, c(1,1,2,2,3))
 })