+++ working/tests/ex1.R
@@ -27,4 +27,4 @@ write-up and comparison with other projects,4,6,write-up")
 start_date = as_date(lubridate::ymd("2018-05-01", tz = "GMT"))
 d$start = start_date + d$start * 7
 d$end = d$start + d$duration * 7
-vistime(d, linewidth = 20)
+gg_vistime(d, linewidth = 20)